chore: rename to responseCode
